requirements.txt
mysql-connector-python~=8.4.0
bcrypt~=3.2.0
Flask~=3.0.3